# Source/CodeGen/v1/TypeScript/VitalServiceApiClient/src/models/AddProccessRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * VitalService
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { ProcessPriorityEnum } from './ProcessPriorityEnum';
import {
    ProcessPriorityEnumFromJSON,
    ProcessPriorityEnumFromJSONTyped,
    ProcessPriorityEnumToJSON,
} from './ProcessPriorityEnum';

/**
 * 
 * @export
 * @interface AddProccessRequest
 */
export interface AddProccessRequest {
    /**
     * 
     * @type {string}
     * @memberof AddProccessRequest
     */
    processName?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AddProccessRequest
     */
    alias?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AddProccessRequest
     */
    executionPath?: string | null;
    /**
     * 
     * @type {ProcessPriorityEnum}
     * @memberof AddProccessRequest
     */
    processPriority?: ProcessPriorityEnum;
    /**
     * 
     * @type {Array<number>}
     * @memberof AddProccessRequest
     */
    affinity?: Array<number> | null;
    /**
     * 
     * @type {number}
     * @memberof AddProccessRequest
     */
    profileId?: number;
}

/**
 * Check if a given object implements the AddProccessRequest interface.
 */
export function instanceOfAddProccessRequest(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function AddProccessRequestFromJSON(json: any): AddProccessRequest {
    return AddProccessRequestFromJSONTyped(json, false);
}

export function AddProccessRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): AddProccessRequest {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'processName': !exists(json, 'processName') ? undefined : json['processName'],
        'alias': !exists(json, 'alias') ? undefined : json['alias'],
        'executionPath': !exists(json, 'executionPath') ? undefined : json['executionPath'],
        'processPriority': !exists(json, 'processPriority') ? undefined : ProcessPriorityEnumFromJSON(json['processPriority']),
        'affinity': !exists(json, 'affinity') ? undefined : json['affinity'],
        'profileId': !exists(json, 'profileId') ? undefined : json['profileId'],
    };
}

export function AddProccessRequestToJSON(value?: AddProccessRequest | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'processName': value.processName,
        'alias': value.alias,
        'executionPath': value.executionPath,
        'processPriority': ProcessPriorityEnumToJSON(value.processPriority),
        'affinity': value.affinity,
        'profileId': value.profileId,
    };
}