# Source/Services/VitalService/VitalService/swagger/v1/VitalService.yaml
openapi: 3.0.1
info:
  title: VitalService
  version: '1.0'
paths:
...
  /api/Profile:
    get:
      tags:
        - Profile
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProfileDto'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProfileDto'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProfileDto'
    post:
      tags:
        - Profile
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProfileRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/CreateProfileRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CreateProfileRequest'
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProfileModel'
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileModel'
            text/json:
              schema:
                $ref: '#/components/schemas/ProfileModel'
    put:
      tags:
        - Profile
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Success
  '/api/Profile/{id}':
    get:
      tags:
        - Profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProfileDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ProfileDto'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                type: integer
                format: int32
            application/json:
              schema:
                type: integer
                format: int32
            text/json:
              schema:
                type: integer
                format: int32
    delete:
      tags:
        - Profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Success
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                type: integer
                format: int32
            application/json:
              schema:
                type: integer
                format: int32
            text/json:
              schema:
                type: integer
                format: int32
  /api/Profile/process:
    post:
      tags:
        - Profile
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddProccessRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/AddProccessRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/AddProccessRequest'
      responses:
        '200':
          description: Success
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                type: integer
                format: int32
            application/json:
              schema:
                type: integer
                format: int32
            text/json:
              schema:
                type: integer
                format: int32
    put:
      tags:
        - Profile
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateManagedRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/UpdateManagedRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/UpdateManagedRequest'
      responses:
        '200':
          description: Success
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                type: integer
                format: int32
            application/json:
              schema:
                type: integer
                format: int32
            text/json:
              schema:
                type: integer
                format: int32
  '/api/Profile/process/{id}':
    delete:
      tags:
        - Profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Success
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                type: integer
                format: int32
            application/json:
              schema:
                type: integer
                format: int32
            text/json:
              schema:
                type: integer
                format: int32
...
components:
  schemas:
    AddProccessRequest:
      type: object
      properties:
        processName:
          type: string
          nullable: true
        alias:
          type: string
          nullable: true
        executionPath:
          type: string
          nullable: true
        processPriority:
          $ref: '#/components/schemas/ProcessPriorityEnum'
        affinity:
          type: array
          items:
            type: integer
            format: int32
          nullable: true
        profileId:
          type: integer
          format: int32
      additionalProperties: false
...
    CreateProfileRequest:
      type: object
      properties:
        name:
          type: string
          nullable: true
      additionalProperties: false
...
    ManagedModel:
      type: object
      properties:
        id:
          type: integer
          format: int32
        processName:
          type: string
          nullable: true
        executionPath:
          type: string
          nullable: true
        alias:
          type: string
          nullable: true
        processPriority:
          $ref: '#/components/schemas/ProcessPriorityEnum'
        affinityBinary:
          type: string
          nullable: true
        parentProfileId:
          type: integer
          format: int32
        profileModel:
          $ref: '#/components/schemas/ProfileModel'
      additionalProperties: false
    ManagedModelDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
        processName:
          type: string
          nullable: true
        alias:
          type: string
          nullable: true
        processPriority:
          $ref: '#/components/schemas/ProcessPriorityEnum'
        affinity:
          type: array
          items:
            type: integer
            format: int32
          nullable: true
        parentProfileId:
          type: integer
          format: int32
      additionalProperties: false
...
    ProcessPriorityEnum:
      enum:
        - DontOverride
        - Idle
        - BelowNormal
        - Normal
        - AboveNormal
        - High
        - RealTime
      type: string
...
    ProfileDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
          nullable: true
        managedModelsIds:
          type: array
          items:
            type: integer
            format: int32
          nullable: true
        enabled:
          type: boolean
        active:
          type: boolean
        priority:
          type: integer
          format: int32
          nullable: true
      additionalProperties: false
    ProfileModel:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
          nullable: true
        managedModels:
          type: array
          items:
            $ref: '#/components/schemas/ManagedModel'
          nullable: true
        enabled:
          type: boolean
        priority:
          type: integer
          format: int32
          nullable: true
      additionalProperties: false
...
    UpdateManagedRequest:
      type: object
      properties:
        managedModelDto:
          $ref: '#/components/schemas/ManagedModelDto'
      additionalProperties: false
    UpdateProfileRequest:
      type: object
      properties:
        profile:
          $ref: '#/components/schemas/ProfileDto'
      additionalProperties: false